+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-**02/09/2020**
+* **02/09/2020**
     + Adding Gaussian low-pass filtering to script ``/code/tutorials/exp_somb/post_rec_particles.py``
 * **26/12/2020**
     + Adding comments to explain XML file
 * **27/12/2020**
-    + Using global_analysis within pre-procesing segmentation scipts in PySeg tutorials
+    + Using global_analysis within pre-procesing segmentation scripts in PySeg tutorials
+* **26/01/2021**
+    + Adding particle surface overlapping fraction option